+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Change: Drop compatibility for version of WP older than 4.9
 * Change: Update links to coffee2code.com to be HTTPS
 * Unit tests:
+    * Fix: Ensure admin-related tests call `test_turn_on_admin()` so admin init actions are called
     * Fix: Invoke parent class's `setUp()` during `setUp()`
     * Change: Remove unnecessary unregistering of hooks in `tearDown()`
     * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests (and delete commented-out code)